Update README.md
Business Problem:
One of the global banks would like to understand what factors driving credit card spend are. The bank want use these insights to calculate credit limit. In order to solve the problem, the bank conducted survey of 5000 customers and collected data.

The objective of this case study is to understand what's driving the total spend (Primary Card + Secondary card). Given the factors, predict credit limit for the new applicants

Data Availability:
Data for the case are available in xlsx format.

The data have been provided for 5000 customers.

Detailed data dictionary has been provided for understanding the data in the data.

Data is encoded in the numerical format to reduce the size of the data however some of the variables are categorical. You can find the details in the data dictionary.

Please find the following files:
1.Required ipynb file

2.Provided data and other material